Fix logger reference in processWarningsAndErrors.
